requirements.txt
# Required dependencies for Date Extractor
PyPDF2>=3.0.0
pillow>=10.0.0
pytesseract>=0.3.10
python-dateutil>=2.8.2